add filters
plugins now handle filtering (kinda)
